refactor(documents-tests): Organize the generated documents that's been sitting around.
