--- backend/TEST_OUTPUT_IMPROVEMENTS.md
@@ -1,188 +1,0 @@
-# Test Output Improvements
-
-## What Was Fixed
-
-### Problem
-Test output was extremely noisy with:
-- ❌ Minified Prisma runtime library code (thousands of lines)
-- ❌ Internal stack traces from node_modules
-- ❌ Application console.error logs from controllers
-- ❌ Verbose debugging information
-- ❌ Hard to identify actual test failures
-
-### Solution
-Implemented intelligent console filtering and custom reporting:
-
-## Changes Made
-
-### 1. Enhanced Console Filtering (`tests/setup.js`)
-**Before**: All console output shown, including minified code
-**After**: Complete suppression of noise, only shows when `SHOW_LOGS=true`
-
-```javascript
-// Now suppresses:
-// - All console.log during tests
-// - All console.warn during tests  
-// - All console.error during tests (errors still visible in test failures)
-// - Minified Prisma library code
-// - Internal stack traces
-```
-
-### 2. Improved Custom Reporter (`tests/customReporter.js`)
-**Before**: Showed full error messages with minified code
-**After**: Intelligently filters error messages
-
-Features:
-- ✅ Removes minified Prisma runtime code
-- ✅ Filters internal stack traces
-- ✅ Shows only relevant error lines
-- ✅ Clean summary at the end
-- ✅ Proper formatting with emoji indicators
-
-Detects and removes:
-- Lines matching `/function \w+\([^)]*\){/` (minified functions)
-- Lines with `let t=` and `let r=` (minified variables)
-- Lines from `src/generated/prisma/runtime/library.js`
-- Lines from `node_modules`
-- Internal Jest stack frames
-
-### 3. Updated Jest Configuration (`jest.config.js`)
-```javascript
-{
-  verbose: false,           // Suppress verbose output
-  silent: false,            // But don't silence everything
-  reporters: [
-    ['default', { 
-      summary: true,
-      summaryThreshold: 0
-    }],
-    '<rootDir>/tests/customReporter.js'  // Custom clean summaries
-  ]
-}
-```
-
-### 4. Added Test Scripts (`package.json`)
-```json
-{
-  "test": "jest --runInBand",              // Clean output (default)
-  "test:verbose": "SHOW_LOGS=true jest --runInBand",  // Full debugging
-  "test:quiet": "jest --silent --runInBand",          // Minimal output
-  "test:summary": "jest --verbose=false --runInBand", // Summary only
-  "test:failures": "jest --onlyFailures --runInBand"  // Rerun failed tests
-}
-```
-
-## How to Use
-
-### Normal Testing (Clean Output)
-```bash
-npm test
-```
-**Shows**: Test progress, failures only, clean summary
-
-### Debugging Mode (Full Logs)
-```bash
-# PowerShell
-$env:SHOW_LOGS="true"; npm test
-
-# Or use the script
-npm run test:verbose
-```
-**Shows**: Everything including application logs
-
-### Minimal Output
-```bash
-npm run test:quiet
-```
-**Shows**: Only final summary
-
-### Re-run Failed Tests
-```bash
-npm run test:failures
-```
-**Shows**: Only tests that failed in the previous run
-
-## Output Comparison
-
-### Before
-```
-[thousands of lines of minified code]
-function rm(e,r,t){let n=t.getComputedFields(),i={...
-let t=e._runtimeDataModel.models[r].fields.reduce((l,u)=>({...
-[more minified code]
-[stack traces from node_modules]
-at Object.<anonymous> (node_modules/jest/...)
-[application errors]
-Error occurred: PrismaClientValidationError...
-List jobs error: ...
-Apply to job error: ...
-[Finally, somewhere: the actual test failure]
-```
-
-### After
-```
-PASS  tests/controllers/authController.test.js
-PASS  tests/controllers/profileController.test.js  
-FAIL  tests/controllers/jobDocumentController.test.js
-
-  ● Job Document Controller › should return 404 when no resume exists
-
-    PrismaClientValidationError:
-    Invalid `prisma.job.create()` invocation
-    
-    Argument `companyName` is missing.
-    
-      186 })
-      187
-      188 // Create a test job
-    → 189 const job = await prisma.job.create({
-
-================================================================================
-📋 FAILED TESTS SUMMARY  
-================================================================================
-
-1. File: /tests/controllers/jobDocumentController.test.js
---------------------------------------------------------------------------------
-   ❌ Job Document Controller › should return 404 when no resume exists
-   
-   PrismaClientValidationError: Argument `companyName` is missing.
-
-================================================================================
-📊 Test Suites: 1 failed, 25 passed, 26 total
-📊 Tests:       2 failed, 199 passed, 201 total
-================================================================================
-
-💡 Tip: Run with SHOW_LOGS=true npm test to see detailed logs
-```
-
-## Benefits
-
-1. **Readable**: See actual test failures immediately
-2. **Fast**: No waiting for thousands of lines to print
-3. **Debuggable**: Enable verbose mode when needed
-4. **Professional**: Clean output suitable for CI/CD
-5. **Flexible**: Multiple output modes for different needs
-
-## For CI/CD
-
-The default `npm test` now produces clean, parseable output perfect for:
-- GitHub Actions
-- GitLab CI
-- Jenkins
-- Any CI system
-
-The custom reporter ensures failed tests are summarized at the end, making it easy to identify issues in CI logs.
-
-## Troubleshooting
-
-If you need to see everything (debugging):
-```bash
-npm run test:verbose
-```
-
-If tests seem to hang, check for:
-- Unclosed database connections
-- Unresolved promises
-- Long-running operations
-
-The `forceExit: true` setting in `jest.config.js` will force exit after tests complete.
