+++ backend/documents/summary/authorization-and-authentication/AUTHORIZATION_EXAMPLE.md
