+++ backend/documents/summary/authorization-and-authentication/STAFF_ADMIN_REGISTRATION.md
