+++ backend/generated-manual-tests/authorization-demo.js
