+++ backend/generated-manual-tests/test-manual.js
@@ -1,5 +1,5 @@
 const request = require('supertest')
-const app = require('./src/app')
+const app = require('../src/app')
 
 async function testBasicEndpoints() {
   console.log('Testing basic endpoints...')
