+++ backend/generated-manual-tests/test-new-registration.js
